changed gaussian type from int to float

+++ myConvolution.h
@@ -11,7 +11,7 @@ enum FILTERING {
 	DERIVATIVE,
 };
 
-uchar calculation1(const Mat kernel, Mat pixels);
+float calculation1(const Mat kernel, Mat pixels);
 float calculation2(const Mat kernel, Mat pixels);
 Mat myConvolution(const Mat src, Mat& kernel, FILTERING flag);
